First round of edits and moves as part of moving the Reporting and Charting code to Academic Observatory. There are more to come but I thought I would make this commit of a manageable size: There is matching set of changes on COKI.
Directory/File / folder	Within group	Class / Function / List	Move action
-------------------------------------------------------------------------------
~/analysis/charts/reports.py	n/a	n/a	Moved reports.py to the new 'reports' directory and changed it to be called report_charts.py

~/analysis/reports	n/a	n/a	Added a new module called 'reports' to Academic Observatory, as per issue #100 Organisation-1

~/analysis/charts/charts.py	class ScatterPlot	class ScatterPlot	Moved this to its own file on COKI/charts called scatterplot.py

~/analysis/charts/charts.py	class ScatterPlot	__init__	Moved this to its own file on COKI/charts called scatterplot.py

~/analysis/charts/charts.py	class ScatterPlot	process_data	Moved this to its own file on COKI/charts called scatterplot.py

~/analysis/charts/charts.py	class ScatterPlot	plot	Moved this to its own file on COKI/charts called scatterplot.py

~/analysis/charts/charts.py	class ScatterPlot	animate	Moved this to its own file on COKI/charts called scatterplot.py

~/analysis/charts/charts.py	class ScatterPlot	anim_frame	Moved this to its own file on COKI/charts called scatterplot.py

~/analysis/charts/rankings.py	class BoxScatter	class BoxScatter	Fixed typo in docstring.

~/analysis/charts/reports.py	class OutputTypesPieChart	class OutputTypesPieChart	This is a duplicate of a function COKI/layouts.py. Moved this to its own file on AcOb/charts called piecharts.py and deleted the COKI/layouts.py one.  I did consider whether only generic piecharts should be in the new piecharts.py but I think these are generic enough for the type of data we have in the Academic Observatory. I am happy to revisit this paradigm

~/analysis/charts/reports.py	class OutputTypesPieChart	__init__	This is a duplicate of a function COKI/layouts.py. Moved this to its own file on AcOb/charts called piecharts.py and deleted the COKI/layouts.py one.  I did consider whether only generic piecharts should be in the new piecharts.py but I think these are generic enough for the type of data we have in the Academic Observatory. I am happy to revisit this paradigm

~/analysis/charts/reports.py	class OutputTypesPieChart	process_data	This is a duplicate of a function COKI/layouts.py. Moved this to its own file on AcOb/charts called piecharts.py and deleted the COKI/layouts.py one.  I did consider whether only generic piecharts should be in the new piecharts.py but I think these are generic enough for the type of data we have in the Academic Observatory. I am happy to revisit this paradigm

~/analysis/charts/reports.py	class OutputTypesPieChart	plot	This is a duplicate of a function COKI/layouts.py. Moved this to its own file on AcOb/charts called piecharts.py and deleted the COKI/layouts.py one.  I did consider whether only generic piecharts should be in the new piecharts.py but I think these are generic enough for the type of data we have in the Academic Observatory. I am happy to revisit this paradigm
