+++ academic_observatory/analysis/reports/report_charts.py
@@ -26,54 +26,6 @@ from academic_observatory.analysis.charts import (
 from academic_observatory.analysis import helpers, defaults
 
 
-class OutputTypesPieChart(AbstractObservatoryChart):
-    """Generate a Pie Graph based on Output Types
-    """
-
-    def __init__(self,
-                 df: pd.DataFrame,
-                 identifier: str,
-                 focus_year: int):
-        self.df = df
-        self.focus_year = focus_year
-        self.identifier = identifier
-
-    def process_data(self):
-        type_categories = defaults.output_types
-        figdata = self.df[(self.df.id == self.identifier) &
-                          (self.df.published_year == self.focus_year) &
-                          (self.df.type.isin(type_categories))
-                          ][['type', 'total']]
-        figdata['type_category'] = pd.Categorical(
-            figdata.type,
-            categories=defaults.output_types,
-            ordered=True)
-        figdata = figdata.set_index('type_category')
-        self.figdata = figdata
-        return self.figdata
-
-    def plot(self, ax=None, **kwargs):
-        if not ax:
-            self.fig, ax = plt.subplots()
-        else:
-            self.fig = ax.get_figure()
-        palette = [defaults.outputs_palette[k] for k in defaults.output_types]
-        outputs_pie = self.figdata.plot.pie(y='total',
-                                            startangle=90,
-                                            labels=None,
-                                            legend=True,
-                                            colors=palette,
-                                            ax=ax)
-        outputs_pie.set_ylabel('')
-        my_circle = plt.Circle((0, 0), 0.4, color='white')
-        p = plt.gcf()
-        p.gca().add_artist(my_circle)
-        outputs_pie.legend(labels=defaults.output_types,
-                           bbox_to_anchor=(1, 0.8))
-        self.fig = outputs_pie.get_figure()
-        return self.fig
-
-
 class GenericTimeChart(AbstractObservatoryChart):
     """Generic Graph of values over time by publication year
     """
